requirements.txt
ultralytics==8.2.2
pillow
matplotlib
scikit-image
torch
polars
opencv-python
tqdm
typing-extensions

# interactive opt
jupyterlab
notebook